+++ app/.pylintrc
@@ -57,7 +57,7 @@ suggestion-mode=yes
 
 # Allow loading of arbitrary C extensions. Extensions are imported into the
 # active Python interpreter and may run arbitrary code.
-unsafe-load-any-extension=yes
+unsafe-load-any-extension=no
 
 
 [MESSAGES CONTROL]
@@ -190,8 +190,7 @@ score=yes
 [REFACTORING]
 
 # Maximum number of nested blocks for function / method body
-#max-nested-blocks=5
-max-nested-blocks=6
+max-nested-blocks=5
 
 # Complete name of functions that never returns. When checking for
 # inconsistent-return-statements if a never returning function is called then
@@ -236,9 +235,7 @@ class-attribute-naming-style=any
 #class-attribute-rgx=
 
 # Naming style matching correct class constant names.
-#class-const-naming-style=UPPER_CASE
-class-const-naming-style=any
-
+class-const-naming-style=UPPER_CASE
 
 # Regular expression matching correct class constant names. Overrides class-
 # const-naming-style.
@@ -345,7 +342,7 @@ indent-string='    '
 max-line-length=100
 
 # Maximum number of lines in a module.
-max-module-lines=1400
+max-module-lines=1000
 
 # Allow the body of a class to be on the same line as the declaration if body
 # contains single statement.
@@ -548,8 +545,7 @@ valid-metaclass-classmethod-first-arg=cls
 [DESIGN]
 
 # Maximum number of arguments for function / method.
-#max-args=11
-max-args=11
+max-args=5
 
 # Maximum number of attributes for a class (see R0902).
 max-attributes=7
@@ -558,30 +554,25 @@ max-attributes=7
 max-bool-expr=5
 
 # Maximum number of branch for function / method body.
-#max-branches=12
-max-branches=30
+max-branches=12
 
 # Maximum number of locals for function / method body.
-#max-locals=15
-max-locals=35
+max-locals=15
 
 # Maximum number of parents for a class (see R0901).
 max-parents=7
 
 # Maximum number of public methods for a class (see R0904).
-#max-public-methods=20
-max-public-methods=30
+max-public-methods=20
 
 # Maximum number of return / yield for function / method body.
 max-returns=6
 
 # Maximum number of statements in function / method body.
-#max-statements=50
-max-statements=65
+max-statements=50
 
 # Minimum number of public methods for a class (see R0903).
-#min-public-methods=2
-min-public-methods=0
+min-public-methods=2
 
 
 [IMPORTS]
