+++ app/schemas_nlp.py
@@ -1,5 +1,5 @@
 '''Defines all input and output classes for translation Apps related API endpoints'''
-
+#pylint: disable=C0103,too-few-public-methods,E0611
 from typing import List, Tuple
 from enum import Enum
 from pydantic import BaseModel, Field, constr, root_validator
